move nested_items_path to ansiblelint.yaml_utils
