+++ src/ansiblelint/rules/ComparisonToEmptyStringRule.py
@@ -7,7 +7,7 @@ from typing import TYPE_CHECKING, Any, Dict, Union
 
 from ansiblelint.rules import AnsibleLintRule
 from ansiblelint.testing import RunFromText
-from ansiblelint.utils import nested_items_path
+from ansiblelint.yaml_utils import nested_items_path
 
 if TYPE_CHECKING:
     from typing import Optional
