+++ src/ansiblelint/rules/ComparisonToLiteralBoolRule.py
@@ -5,7 +5,7 @@ import re
 from typing import TYPE_CHECKING, Any, Dict, Union
 
 from ansiblelint.rules import AnsibleLintRule
-from ansiblelint.utils import nested_items_path
+from ansiblelint.yaml_utils import nested_items_path
 
 if TYPE_CHECKING:
     from typing import Optional
