+++ src/ansiblelint/rules/NoTabsRule.py
@@ -4,7 +4,7 @@ import sys
 from typing import TYPE_CHECKING, Any, Dict, Union
 
 from ansiblelint.rules import AnsibleLintRule
-from ansiblelint.utils import nested_items_path
+from ansiblelint.yaml_utils import nested_items_path
 
 if TYPE_CHECKING:
     from typing import Optional
